requirements.txt
flask
python-dotenv>=0.19.2
authlib>=1.0
requests>=2.27.1